rename wrapper to nvgwrapper
