+++ library/nanovg/Makefile
@@ -150,7 +150,7 @@ endif
 #---------------------------------------------------------------------------------
 clean:
 	@echo clean ...
-	@rm -fr build lib
+	@rm -fr build data lib
 
 #---------------------------------------------------------------------------------
 else
